+++ make.lua
@@ -20,7 +20,7 @@ local includeCodeFormat = true
 
 require "make.detect_platform"
 
-lm:import "3rd/bee.lua"
+lm:import "3rd/bee.lua/make.lua"
 lm:import "make/code_format.lua"
 
 lm:source_set 'lpeglabel' {
@@ -47,9 +47,11 @@ lm:source_set 'luafilesystem' {
 
 lm:executable "lua-language-server" {
     deps = {
+        "source_bee",
+        "source_lua",
+        "source_bootstrap",
         "lpeglabel",
         "luafilesystem",
-        "source_bootstrap",
         includeCodeFormat and "code_format" or nil,
     },
     includes = {
@@ -71,21 +73,21 @@ lm:executable "lua-language-server" {
 }
 
 local platform = require 'bee.platform'
-local exe      = platform.OS == 'Windows' and ".exe" or ""
+local exe      = platform.os == 'windows' and ".exe" or ""
 
 lm:copy "copy_lua-language-server" {
-    input = lm.bindir .. "/lua-language-server" .. exe,
-    output = "bin/lua-language-server" .. exe,
+    inputs = "$bin/lua-language-server" .. exe,
+    outputs = "bin/lua-language-server" .. exe,
 }
 
 lm:copy "copy_bootstrap" {
-    input = "make/bootstrap.lua",
-    output = "bin/main.lua",
+    inputs = "make/bootstrap.lua",
+    outputs = "bin/main.lua",
 }
 
 lm:msvc_copydll 'copy_vcrt' {
     type = "vcrt",
-    output = "bin",
+    outputs = "bin",
 }
 
 lm:phony "all" {
@@ -109,13 +111,13 @@ if lm.notest then
 end
 
 lm:rule "run-bee-test" {
-    lm.bindir .. "/lua-language-server" .. exe, "$in",
+    args = { "$bin/lua-language-server" .. exe, "$in" },
     description = "Run test: $in.",
     pool = "console",
 }
 
 lm:rule "run-unit-test" {
-    "bin/lua-language-server" .. exe, "$in",
+    args = { "bin/lua-language-server" .. exe, "$in" },
     description = "Run test: $in.",
     pool = "console",
 }
@@ -123,13 +125,13 @@ lm:rule "run-unit-test" {
 lm:build "bee-test" {
     rule = "run-bee-test",
     deps = { "lua-language-server", "copy_script" },
-    input = "3rd/bee.lua/test/test.lua",
+    inputs = "3rd/bee.lua/test/test.lua",
 }
 
 lm:build 'unit-test' {
     rule = "run-unit-test",
     deps = { "bee-test", "all" },
-    input = "test.lua",
+    inputs = "test.lua",
 }
 
 lm:default {