+++ script/cli/init.lua
@@ -8,7 +8,17 @@ if _G['CHECK'] then
     os.exit(0, true)
 end
 
+if _G['DOC_UPDATE'] then
+    require 'cli.doc' .runCLI()
+    os.exit(0, true)
+end
+
 if _G['DOC'] then
     require 'cli.doc' .runCLI()
     os.exit(0, true)
+end
+
+if _G['VISUALIZE'] then
+	local ret = require 'cli.visualize' .runCLI()
+	os.exit(ret or 0, true)
 end