+++ script/config/template.lua
@@ -223,8 +223,8 @@ local template = {
                                                 '||', '&&', '!', '!=',
                                                 'continue',
                                             }),
-    ['Lua.runtime.plugin']                  = Type.String,
-    ['Lua.runtime.pluginArgs']              = Type.Array(Type.String),
+    ['Lua.runtime.plugin']                  = Type.Or(Type.String, Type.Array(Type.String)) ,
+    ['Lua.runtime.pluginArgs']              = Type.Or(Type.Array(Type.String), Type.Hash(Type.String, Type.String)),
     ['Lua.runtime.fileEncoding']            = Type.String >> 'utf8' << {
                                                 'utf8',
                                                 'ansi',
@@ -317,7 +317,12 @@ local template = {
     ['Lua.workspace.maxPreload']            = Type.Integer >> 5000,
     ['Lua.workspace.preloadFileSize']       = Type.Integer >> 500,
     ['Lua.workspace.library']               = Type.Array(Type.String),
-    ['Lua.workspace.checkThirdParty']       = Type.Boolean >> true,
+    ['Lua.workspace.checkThirdParty']       = Type.Or(Type.String >> 'Ask' << {
+                                                'Ask',
+                                                'Apply',
+                                                'ApplyInMemory',
+                                                'Disable',
+                                            }, Type.Boolean),
     ['Lua.workspace.userThirdParty']        = Type.Array(Type.String),
     ['Lua.completion.enable']               = Type.Boolean >> true,
     ['Lua.completion.callSnippet']          = Type.String  >> 'Disable' << {
@@ -392,6 +397,7 @@ local template = {
     ['Lua.type.castNumberToInteger']        = Type.Boolean >> true,
     ['Lua.type.weakUnionCheck']             = Type.Boolean >> false,
     ['Lua.type.weakNilCheck']               = Type.Boolean >> false,
+    ['Lua.type.inferParamType']            = Type.Boolean >> false,
     ['Lua.doc.privateName']                 = Type.Array(Type.String),
     ['Lua.doc.protectedName']               = Type.Array(Type.String),
     ['Lua.doc.packageName']                 = Type.Array(Type.String),