+++ script/files.lua
@@ -78,7 +78,7 @@ end
 ---@param uri uri
 ---@return uri
 function m.getRealUri(uri)
-    if platform.OS ~= 'Windows' then
+    if platform.os ~= 'windows' then
         return furi.normalize(uri)
     end
     if not furi.isValid(uri) then
@@ -655,6 +655,17 @@ function m.compileStateAsync(uri, callback)
     end)
 end
 
+local function pluginOnTransformAst(uri, state)
+    local plugin   = require 'plugin'
+    ---TODO: maybe deepcopy astNode
+    local suc, result = plugin.dispatch('OnTransformAst', uri, state.ast)
+    if not suc then
+        return state
+    end
+    state.ast = result or state.ast
+    return state
+end
+
 ---@param uri uri
 ---@return parser.state?
 function m.compileState(uri)
@@ -697,6 +708,12 @@ function m.compileState(uri)
 
     if not state then
         log.error('Compile failed:', uri, err)
+        return nil
+    end
+
+    state = pluginOnTransformAst(uri, state)
+    if not state then
+        log.error('pluginOnTransformAst failed! discard the file state')
         return nil
     end
 
@@ -816,7 +833,7 @@ function m.isDll(uri)
     if not ext then
         return false
     end
-    if platform.OS == 'Windows' then
+    if platform.os == 'windows' then
         if ext == 'dll' then
             return true
         end
@@ -915,7 +932,7 @@ function m.normalize(path)
             break
         end
     end
-    if platform.OS == 'Windows' then
+    if platform.os == 'windows' then
         path = path:gsub('[/\\]+', '\\')
                    :gsub('[/\\]+$', '')
                    :gsub('^(%a:)$', '%1\\')