+++ script/parser/luadoc.lua
@@ -157,6 +157,7 @@ Symbol              <-  ({} {
 ---@field generics?         parser.object[]
 ---@field generic?          parser.object
 ---@field docAttr?          parser.object
+---@field pattern?          string
 
 local function parseTokens(text, offset)
     Ci = 0
@@ -362,6 +363,78 @@ local function parseTable(parent)
             nextToken()
         else
             nextSymbolOrError('}')
+            break
+        end
+    end
+    typeUnit.finish = getFinish()
+    return typeUnit
+end
+
+local function parseTuple(parent)
+    if not checkToken('symbol', '[', 1) then
+        return nil
+    end
+    nextToken()
+    local typeUnit = {
+        type    = 'doc.type.table',
+        start   = getStart(),
+        parent  = parent,
+        fields  = {},
+        isTuple = true,
+    }
+
+    local index = 1
+    while true do
+        if checkToken('symbol', ']', 1) then
+            nextToken()
+            break
+        end
+        local field = {
+            type   = 'doc.type.field',
+            parent = typeUnit,
+        }
+
+        do
+            local needCloseParen
+            if checkToken('symbol', '(', 1) then
+                nextToken()
+                needCloseParen = true
+            end
+            field.name = {
+                type        = 'doc.type',
+                start       = getFinish(),
+                firstFinish = getFinish(),
+                finish      = getFinish(),
+                parent      = field,
+            }
+            field.name.types = {
+                [1] = {
+                    type   = 'doc.type.integer',
+                    start  = getFinish(),
+                    finish = getFinish(),
+                    parent = field.name,
+                    [1]    = index,
+                }
+            }
+            index          = index + 1
+            field.extends  = parseType(field)
+            if not field.extends then
+                break
+            end
+            field.optional = field.extends.optional
+            field.start    = field.extends.start
+            field.finish   = field.extends.finish
+            if needCloseParen then
+                nextSymbolOrError ')'
+            end
+        end
+
+        typeUnit.fields[#typeUnit.fields+1] = field
+        if checkToken('symbol', ',', 1)
+        or checkToken('symbol', ';', 1) then
+            nextToken()
+        else
+            nextSymbolOrError(']')
             break
         end
     end
@@ -633,6 +706,53 @@ local function parseCode(parent)
     return code
 end
 
+local function parseCodePattern(parent)
+    local tp, pattern = peekToken()
+    if not tp or tp ~= 'name' then
+        return nil
+    end
+    local codeOffset
+    local finishOffset
+    local content
+    for i = 2, 8 do
+        local next, nextContent = peekToken(i)
+        if not next or TokenFinishs[Ci+i-1] + 1 ~= TokenStarts[Ci+i] then
+            if codeOffset then
+                finishOffset = i
+                break
+            end
+            ---不连续的name，无效的
+            return nil
+        end
+        if next == 'code' then
+            if codeOffset and content ~= nextContent then
+                -- 暂时不支持多generic
+                return nil
+            end
+            codeOffset = i
+            pattern = pattern .. "%s"
+            content = nextContent
+        elseif next ~= 'name' then
+            return nil
+        else
+            pattern = pattern .. nextContent
+        end
+    end
+    local start = getStart()
+    for i = 2 , finishOffset do
+        nextToken()
+    end
+    local code = {
+        type   = 'doc.type.code',
+        start  = start,
+        finish = getFinish(),
+        parent = parent,
+        pattern = pattern,
+        [1]    = content,
+    }
+    return code
+end
+
 local function parseInteger(parent)
     local tp, content = peekToken()
     if not tp or tp ~= 'integer' then
@@ -682,11 +802,13 @@ end
 function parseTypeUnit(parent)
     local result = parseFunction(parent)
                 or parseTable(parent)
+                or parseTuple(parent)
                 or parseString(parent)
                 or parseCode(parent)
                 or parseInteger(parent)
                 or parseBoolean(parent)
                 or parseParen(parent)
+                or parseCodePattern(parent)
     if not result then
         result = parseName('doc.type.name', parent)
               or parseDots('doc.type.name', parent)
@@ -864,6 +986,7 @@ local docSwitch = util.switch()
         while true do
             local extend = parseName('doc.extends.name', result)
                         or parseTable(result)
+                        or parseTuple(result)
             if not extend then
                 pushWarning {
                     type   = 'LUADOC_MISS_CLASS_EXTENDS_NAME',
@@ -903,6 +1026,7 @@ local docSwitch = util.switch()
         local result = {
             type   = 'doc.alias',
         }
+        result.docAttr = parseDocAttr(result)
         result.alias = parseName('doc.alias.name', result)
         if not result.alias then
             pushWarning {
@@ -1546,7 +1670,7 @@ local function trimTailComment(text)
     and comment:find '[\'"%]]%s*$' then
         local state = compile(comment:gsub('^%s+', ''), 'String')
         if state and state.ast then
-            comment = state.ast[1]
+            comment = state.ast[1] --[[@as string]]
         end
     end
     return util.trim(comment)
@@ -1964,6 +2088,14 @@ local function bindDocWithSources(sources, binded)
     bindGeneric(binded)
     bindCommentsAndFields(binded)
     bindReturnIndex(binded)
+    
+    -- doc is special node
+    if lastDoc.special then
+        if bindDoc(lastDoc.special, binded) then
+            return
+        end
+    end
+
     local row = guide.rowColOf(lastDoc.finish)
     local suc = bindDocsBetween(sources, binded, guide.positionOf(row, 0), lastDoc.start)
     if not suc then
@@ -1994,12 +2126,16 @@ local function bindDocs(state)
             state.ast.docs.groups[#state.ast.docs.groups+1] = binded
         end
         binded[#binded+1] = doc
-        if isTailComment(text, doc) then
+		if doc.specialBindGroup then
+			bindDocWithSources(sources, doc.specialBindGroup)
+			binded = nil
+		elseif isTailComment(text, doc) and doc.type ~= "doc.class" and doc.type ~= "doc.field" then
             bindDocWithSources(sources, binded)
             binded = nil
         else
             local nextDoc = state.ast.docs[i+1]
-            if not isNextLine(doc, nextDoc) then
+            if nextDoc and nextDoc.special
+            or not isNextLine(doc, nextDoc) then
                 bindDocWithSources(sources, binded)
                 binded = nil
             end
@@ -2028,7 +2164,7 @@ local function findTouch(state, doc)
     end
 end
 
-return function (state)
+local function luadoc(state)
     local ast = state.ast
     local comments = state.comms
     table.sort(comments, function (a, b)
@@ -2098,6 +2234,18 @@ return function (state)
         end
     end
 
+    if ast.state.pluginDocs then
+        for i, doc in ipairs(ast.state.pluginDocs) do
+            insertDoc(doc, doc.originalComment)
+        end
+        ---@param a unknown
+        ---@param b unknown
+        table.sort(ast.docs, function (a, b)
+            return a.start < b.start
+        end)
+        ast.state.pluginDocs = nil
+    end
+
     ast.docs.start  = ast.start
     ast.docs.finish = ast.finish
 
@@ -2107,3 +2255,21 @@ return function (state)
 
     bindDocs(state)
 end
+
+return {
+    buildAndBindDoc = function (ast, src, comment, group)
+        local doc = buildLuaDoc(comment)
+        if doc then
+            local pluginDocs = ast.state.pluginDocs or {}
+            pluginDocs[#pluginDocs+1] = doc
+            doc.special = src
+            doc.originalComment = comment
+            doc.virtual = true
+			doc.specialBindGroup = group
+            ast.state.pluginDocs = pluginDocs
+            return doc
+        end
+        return nil
+    end,
+    luadoc = luadoc
+}