+++ test/completion/common.lua
@@ -3864,6 +3864,12 @@ local x = function (x, y) end
 (EXISTS)
 
 TEST [[
+--- <??>
+local x = function (x, y) end
+]]
+(EXISTS)
+
+TEST [[
 local x = {
 <??>
 })
@@ -4387,3 +4393,44 @@ f {
         kind  = define.CompletionItemKind.Property,
     },
 }
+
+TEST [[
+---@class A
+---@overload fun(x: {id: string})
+
+---@generic T
+---@param t `T`
+---@return T
+local function new(t) end
+
+new 'A' {
+    <??>
+}
+]]
+{
+    {
+        label = 'id',
+        kind  = define.CompletionItemKind.Property,
+    }
+}
+
+TEST [[
+---@class namespace.A
+---@overload fun(x: {id: string})
+
+---@generic T
+---@param t namespace.`T`
+---@return T
+local function new(t) end
+
+new 'A' {
+    <??>
+}
+]]
+{
+    {
+        label = 'id',
+        kind  = define.CompletionItemKind.Property,
+    }
+}
+