+++ test/type_inference/init.lua
@@ -4335,4 +4335,11 @@ end
 
 local <?x?> = f()
 ]=]
---
+
+TEST 'boolean|number' [[
+---@alias A number
+---@alias(partial) A boolean
+
+---@type A
+local <?x?>
+]]